fixed italic&bold whitespace bugs
