+++ cs/Markdown/MarkdownToHtmlConverterTests/BoldItalicTestCases.cs
@@ -21,8 +21,8 @@ namespace MarkdownToHtmlConverterTests
             ("Process {0} at borders of different words", "{2}first second{2}", "<{1}>first second</{1}>"),
             ("Process {0} in start of word", "{2}s{2}tart", "<{1}>s</{1}>tart"),
             ("Process {0} at end of word", "en{2}d{2}", "en<{1}>d</{1}>"),
-            ("Ignore open {0} before Space char", "ignore{2} before{2} whitespace", "ignore{2} before{2} whitespace"),
-            ("Ignore closing {0} after Space char", "ignore {2}after {2}whitespace", "ignore {2}after {2}whitespace"),
+            ("Ignore open {0} before Space char", "{2} b{2}c", "{2} b{2}c"),
+            ("Ignore closing {0} after Space char", "{2}bc {2}", "{2}bc {2}"),
             ("Ignore empty {0}", "{2}{2}", "{2}{2}"),
             ("Opening {0} can be screened", @"\{2}abc{2}", "{2}abc{2}"),
             ("Closing {0} can be screened", @"{2}abc\{2}", "{2}abc{2}"),
